fix e_notice, merge changes

+++ CloudVeilManager/app/Http/Middleware/CheckAndUpdateDeviceId.php
@@ -58,12 +58,16 @@ class CheckAndUpdateDeviceId
                 }
             }
             if(!$activation) {
-                $activation = AppUserActivation::withTrashed()->where("identifier", $input['identifier'])->orderBy("last_sync_time", "DESC")->first();
-                if ($activation) {
+                $activation = AppUserActivation::withTrashed()->
+                where("identifier", $input['identifier'])->
+                where("device_id", $input['device_id'])->
+                orderBy("last_sync_time", "DESC")->first();
+                if ($activation && $activation->trashed()) {
                     $activation->restore();
                     Cache::delete('AppUserActivation' . $activationId);
                 }
             }
+
             if ($activation) {
                 if ($appVersion != "") {
                     $activation->app_version = $appVersion;
